refactor(frontend): 提取 EventBus 类到独立文件
将 EventBus 类从 websocket.ts 提取到独立的 event-bus.ts 文件，
遵循单一职责原则，与后端架构保持一致。

变更内容：
- 创建 apps/frontend/src/services/event-bus.ts
- 将 EventBus 类迁移为泛型类，支持不同事件类型映射
- 更新 websocket.ts 导入并使用泛型 EventBus<EventBusEvents>
- 重新导出 EventBus 和 EventListener 类型，保持向后兼容

Closes #3214
